Fix Shaky Sidepanel (#592)
