+++ simplq/src/components/common/QueueStats/QueueStats.jsx
@@ -1,5 +1,9 @@
-import React, { useMemo } from 'react';
+import React, { useEffect, useCallback, useMemo } from 'react';
 import moment from 'moment';
+import { useGetQueueStatus } from 'store/asyncActions';
+import { selectQueueStatus } from 'store/queueStatus';
+import { useDispatch, useSelector } from 'react-redux';
+import { selectTokens } from 'store/selectedQueue';
 import styles from './QueueStats.module.scss';
 
 const DetailRow = ({ title, value, large }) => (
@@ -11,13 +15,20 @@ const DetailRow = ({ title, value, large }) => (
   </div>
 );
 
-export default (props) => {
-  const {
-    status,
-    queueCreationTimestamp,
-    numberOfActiveTokens,
-    totalNumberOfTokens,
-  } = props.queueStatus;
+export default ({ queueId }) => {
+  const dispatch = useDispatch();
+  const getQueueStatus = useCallback(useGetQueueStatus(), []);
+  const tokens = useSelector(selectTokens);
+
+  useEffect(() => {
+    if (queueId) {
+      dispatch(getQueueStatus({ queueId }));
+    }
+  }, [queueId, tokens, dispatch, getQueueStatus]);
+
+  const { status, queueCreationTimestamp, numberOfActiveTokens, totalNumberOfTokens } = useSelector(
+    selectQueueStatus
+  );
 
   const creationTime = useMemo(() => {
     if (!queueCreationTimestamp) return '';
