+++ simplq/src/components/pages/Admin/QueueDetails.jsx
@@ -1,22 +1,9 @@
-import React, { useEffect, useCallback } from 'react';
+import React from 'react';
 import InfoIcon from '@material-ui/icons/Info';
 import QueueStats from 'components/common/QueueStats';
-import { useGetQueueStatus } from 'store/asyncActions';
-import { selectQueueStatus } from 'store/queueStatus';
-import { useDispatch, useSelector } from 'react-redux';
-import { selectTokens } from 'store/selectedQueue';
 import SidePanelItem from 'components/common/SidePanel/SidePanelItem';
 
 export default ({ queueId }) => {
-  const queueStatus = useSelector(selectQueueStatus);
-  const dispatch = useDispatch();
-  const getQueueStatus = useCallback(useGetQueueStatus(), []);
-  const tokens = useSelector(selectTokens);
-
-  useEffect(() => {
-    dispatch(getQueueStatus({ queueId }));
-  }, [queueId, tokens, dispatch, getQueueStatus]);
-
   return (
     <SidePanelItem
       Icon={InfoIcon}
@@ -24,7 +11,7 @@ export default ({ queueId }) => {
       description="Other information about the queue"
       expandable
     >
-      <QueueStats queueStatus={queueStatus} />
+      <QueueStats queueId={queueId} />
     </SidePanelItem>
   );
 };
